+++ code/packages/types/src/types.ts
@@ -308,7 +308,7 @@ export type WeaveNodeConfiguration = {
 // Image api
 
 export type ImageCrossOrigin = 'anonymous' | 'use-credentials';
-export type URLTransformerFunction = (url: string) => string;
+export type URLTransformerFunction = (url: string, node?: Konva.Node) => string;
 
 // Measurement